requirements.txt
flask
flask-cors
pillow
pix2text